+++ shared/src/model/config/trakt.rs
@@ -1,13 +1,8 @@
 use serde::{Deserialize, Serialize};
 use std::fmt;
 use std::str::FromStr;
-
-const  TRAKT_API_KEY: &str = "0183a05ad97098d87287fe46da4ae286f434f32e8e951caad4cc147c947d79a3";
-const  TRAKT_API_VERSION: &str = "2";
-const  TRAKT_API_URL: &str = "https://api.trakt.tv";
-fn default_fuzzy_threshold() -> u8 {
-    80
-}
+use crate::model::DEFAULT_USER_AGENT;
+use crate::utils::{TRAKT_API_KEY, TRAKT_API_URL, TRAKT_API_VERSION, default_trakt_fuzzy_threshold};
 
 #[derive(Debug, Default, Copy, Clone, Serialize, Deserialize, PartialEq, Eq)]
 #[serde(rename_all = "lowercase")]
@@ -45,22 +40,26 @@ impl FromStr for TraktContentType {
 #[derive(Default, Debug, Clone, Serialize, Deserialize, PartialEq)]
 #[serde(deny_unknown_fields)]
 pub struct TraktApiConfigDto {
-    #[serde(default)]
-    pub key: String,
+    #[serde(default, alias = "key")]
+    pub api_key: String,
     #[serde(default)]
     pub version: String,
     #[serde(default)]
     pub url: String,
+    #[serde(default)]
+    pub user_agent: String,
 }
 
 impl TraktApiConfigDto {
     pub fn prepare(&mut self) {
-        let key  =  self.key.trim();
-        self.key = String::from(if key.is_empty() { TRAKT_API_KEY } else { key });
+        let key  =  self.api_key.trim();
+        self.api_key = String::from(if key.is_empty() { TRAKT_API_KEY } else { key });
         let version = self.version.trim();
         self.version = String::from(if version.is_empty() { TRAKT_API_VERSION } else { version });
         let url = self.url.trim();
         self.url = String::from(if url.is_empty() { TRAKT_API_URL } else { url });
+        let user_agent = self.user_agent.trim();
+        self.user_agent = String::from(if user_agent.is_empty() { DEFAULT_USER_AGENT } else { user_agent });
     }
 }
 
@@ -72,7 +71,7 @@ pub struct TraktListConfigDto {
     pub list_slug: String,
     pub category_name: String,
     pub content_type: TraktContentType,
-    #[serde(default = "default_fuzzy_threshold")]
+    #[serde(default = "default_trakt_fuzzy_threshold")]
     pub fuzzy_match_threshold: u8, // Percentage (0-100)
 }
 
@@ -83,7 +82,7 @@ impl Default for TraktListConfigDto {
             list_slug: String::new(),
             category_name: String::new(),
             content_type: TraktContentType::default(),
-            fuzzy_match_threshold: default_fuzzy_threshold(),
+            fuzzy_match_threshold: default_trakt_fuzzy_threshold(),
         }
     }
 }